This makes the app look so much better on ICS
Given GitHub site colours and the early mobile-views I've seen, I think
it's ineveitable that we won't want to use the default 'dark' theme used by
Android and go for the light alternative both pre- and post- Honeycomb.

This little bit of config just ensures that we use the light *holo* theme
on ICS, switches on the actionbar, and just looks much better.