Make sure proguard knows about the concise newArraylist() from Guava
I wanted to use the concise no-generic-declaration newArraylist() method
from Guava in the integration tests, but because the Guava library is
used in the main app, and proguard didn't see the newArraylist() method
there, it stripped it out and the integration test failed with a
'unable to resolve static method' dalvik VM error.

Rather than futz about with a proguard rule, this commit just ensures
that the method is seen in the main app, so proguard doesn't strip it
out.