Clear stored page number when new org. is selected
The lazy loading news list would stop functioning after
an org. was selected since not all the state was properly
cleared from the resource pager.

This also required clearing all fragments from the pager
adapter when the default user to org. context switch
occurs in order to preserve the correct list adapter and
loading footer view that would have been cleared by the
added call to onViewCreated on the fragment.