Put list item update logic in list adapter extensions
Storing details on updating the view inside the holders
presents complexity since holders have to do the bookkeeping
and that often requires extra objects being stored and logic
being run once per view holder.

Moving this to the adapter allows it to be computed once
and reset when the items change and doesn't require storing
state between all the holders.

The new pattern is to create an ItemView extension that exposes
public final fields for each view element that get updated and
then the list adapter will contain the logic on how the view
should be updated.